Add switch to override relative symbols
Add a switch to override discovery of relative offsets in the kallsyms
table.  If the binary under analysis has an absolute base in the upper
half of the virtual address range, all the addresses appear to be
negative numbers.  For 64-bit systems a 50-50 split would put kernel
addresses at 0x80000000-00000000 and for 32 bit systems with a 3G/1G
split, the kernel addresses start at 0xC0000000