Add heuristics to suggest absolute addresses
When looking for relative addresses, perform heuristic checks for
absolute addresses.  The original heuristic was just looking for
negative numbers to be the offsets.  This is insufficient because a
kernel virtual address for a kernel loaded at an absolute address will
almost certainly be in the top half of the virtual address space which
would be a negative number.

2 heuristics have been added.  The first is to check the top 3 nybbles
are 0xFFF.  True negative numbers are unlikely to be *THAT* negative.
The kernel will be on the order of a few 10s of MB

The second heuristic is to check for zeros in the top byte using the
mask 0x3F.  This assumes the kernel is loaded near the bottom of the
kernel address space and will catch the 3G/1G split.  Strictly speaking
the second heuristic should never trip if the first one doesn't.